Split up main.rs and add a test for cleaner module
